fix: [Interceptors] Add a copy button (full endpoint for container source) for Completion endpoint (Optional) fields
