+++ apps/ai-dial-admin/src/components/Assets/Files/View.tsx
@@ -16,7 +16,8 @@ import { useI18n } from '@/src/locales/client';
 import { DialFile } from '@/src/models/dial/file';
 import { ApplicationRoute } from '@/src/types/routes';
 import { getNameExtensionFromFile } from '@/src/utils/files/get-extension';
-import { addTrailingSlash, changePath } from '@/src/utils/files/path';
+import { changePath } from '@/src/utils/files/path';
+import { addTrailingSlash } from '@/src/utils/url';
 import { isEqualSkippingUndefined } from '@/src/utils/is-equals-entity';
 import { getUrnForEntity } from '@/src/utils/open-in-new-tab';
 import FileProperties from './Properties';
