+++ apps/ai-dial-admin/src/components/Common/FolderList/FolderList.tsx
@@ -23,7 +23,8 @@ import { DialFile } from '@/src/models/dial/file';
 import { DialFolder } from '@/src/models/dial/folder';
 import { ApplicationRoute } from '@/src/types/routes';
 import { getFolderName } from '@/src/utils/files/folder';
-import { addTrailingSlash, getFolderNameAndPath, isFolder } from '@/src/utils/files/path';
+import { addTrailingSlash } from '@/src/utils/url';
+import { getFolderNameAndPath, isFolder } from '@/src/utils/files/path';
 import { isAssetView } from '@/src/utils/is-asset-view';
 import FolderListModals, { ModalType } from './Modals/FolderListModals';
 import { generateFolderListFromBulkPaths } from './utils';
