+++ apps/ai-dial-admin/src/components/SourceField/Endpoints/ModelEndpoint.tsx
@@ -1,4 +1,4 @@
-import { FC, useCallback, useEffect, useState } from 'react';
+import { FC, useCallback, useEffect, useMemo, useState } from 'react';
 
 import { DialRadioGroup, RadioGroupOrientation, RadioButtonWithContent } from '@epam/ai-dial-ui-kit';
 
@@ -11,6 +11,7 @@ import { useI18n } from '@/src/locales/client';
 import ComplexInput from '@/src/components/Common/ComplexInput/ComplexInput';
 import { getEndpointPostfix } from '@/src/utils/models/model-endpoint';
 import classNames from 'classnames';
+import { addTrailingSlash } from '@/src/utils/url';
 
 interface Props {
   entity: DialModel;
@@ -30,19 +31,9 @@ const ModelEndpoint: FC<Props> = ({ entity, prefix, onChange, isModal }) => {
   const [postfix, setPostfix] = useState('');
   const [name, setName] = useState('');
   const [endpointError, setEndpointError] = useState<FieldError | null>(null);
-  const [fullValue, setFullValue] = useState('');
-
-  useEffect(() => {
-    if (prefix) {
-      // NOTE: due to not required id part, ensure no double slashes in path
-      const path = entity.source?.completionEndpointPath;
-      const clearedPath = path?.startsWith('/') ? path.slice(1) : path;
-
-      setFullValue(`${prefix}${clearedPath}`);
-    } else {
-      setFullValue(entity.endpoint as string);
-    }
-  }, [entity, prefix]);
+  const fullValue = useMemo(() => {
+    return prefix ? `${addTrailingSlash(prefix)}${entity.endpoint}` : entity.endpoint || '';
+  }, [entity.endpoint, prefix]);
 
   const onChangePath = useCallback(
     (value?: string) => {
