+++ apps/ai-dial-admin/src/utils/tests/urls.spec.ts
@@ -1,0 +1,13 @@
+import { addTrailingSlash } from '../url';
+
+describe('Utils :: addTrailingSlash', () => {
+  test('Should add slash', () => {
+    const res = addTrailingSlash('folder1/folder2/all/folder');
+    expect(res).toEqual('folder1/folder2/all/folder/');
+  });
+
+  test('Should not add slash if exists', () => {
+    const res = addTrailingSlash('folder1/folder2/all/folder/');
+    expect(res).toEqual('folder1/folder2/all/folder/');
+  });
+});
