+++ apps/ai-dial-admin/src/utils/url.ts
@@ -1,0 +1,3 @@
+export const addTrailingSlash = (path: string) => {
+  return path.endsWith('/') ? path : `${path}/`;
+};
